Bypass Post Sign-In Notification
Our company utilizes the Post Sign-In Notification from time to time
to inform us of updates/bugs/etc.  When that's used the scripting would
not be able to finish getting the DSID.  This commit adds code to check
if the Post Sign-In page comes up instead of the expected page.  If so
it posts to the correct page with the necessary form values to get
to the expected page and finish getting the DSID value.